requirements.txt
langchain==0.3.9
scikit-learn==1.3.0  
faiss-cpu==1.7.4
openai==0.29.0
pandas==2.2.0
PyPDF2==3.0.1
#python==3.13.0